+++ app/queries.py
@@ -1,7 +1,7 @@
 from suite.database import Session
 
 from app.cache import region
-from app.models import Currency, ChatRequests
+from app.models import Currency, ChatRequests, Chat
 
 
 @region.cache_on_arguments(expiration_time=300)
@@ -15,9 +15,21 @@ def get_all_currencies():
     return Session.query(Currency.code, Currency.name).filter_by(is_active=True).order_by(Currency.name).all()
 
 
+def get_keyboard_size(chat_id):
+    chat = Session.query(Chat.keyboard_size).filter_by(id=chat_id).first()
+    w, h = chat[0].split('x')
+    return int(w) * int(h)
+
+
 def get_last_request(chat_id):
+    size = get_keyboard_size(chat_id)
+
     return Session.query(ChatRequests).filter_by(
         chat_id=chat_id
     ).order_by(
         ChatRequests.times.desc()
-    ).limit(9).all()
+    ).limit(size).all()
+
+
+def have_last_request(chat_id):
+    return Session.query(ChatRequests).filter_by(chat_id=chat_id).first()